+++ test/with_dummyserver/test_https.py
@@ -36,6 +36,7 @@ from urllib3.exceptions import (
     InsecureRequestWarning,
     SystemTimeWarning,
     InsecurePlatformWarning,
+    MaxRetryError,
 )
 from urllib3.packages import six
 from urllib3.util.timeout import Timeout
@@ -70,6 +71,7 @@ class TestHTTPS(HTTPSDummyServerTestCase):
         https_pool = HTTPSConnectionPool(self.host, self.port,
                                          cert_reqs='CERT_REQUIRED',
                                          ca_certs=DEFAULT_CA)
+        self.addCleanup(https_pool.close)
 
         conn = https_pool._new_conn()
         self.assertEqual(conn.__class__, VerifiedHTTPSConnection)
@@ -159,14 +161,7 @@ class TestHTTPS(HTTPSDummyServerTestCase):
         with mock.patch('warnings.warn') as warn:
             r = https_pool.request('GET', '/')
             self.assertEqual(r.status, 200)
-
-            if sys.version_info >= (2, 7, 9):
-                self.assertFalse(warn.called, warn.call_args_list)
-            else:
-                self.assertTrue(warn.called)
-                call, = warn.call_args_list
-                error = call[0][1]
-                self.assertEqual(error, InsecurePlatformWarning)
+            self.assertFalse(warn.called, warn.call_args_list)
 
     def test_invalid_common_name(self):
         https_pool = HTTPSConnectionPool('127.0.0.1', self.port,
@@ -177,10 +172,11 @@ class TestHTTPS(HTTPSDummyServerTestCase):
         try:
             https_pool.request('GET', '/')
             self.fail("Didn't raise SSL invalid common name")
-        except SSLError as e:
+        except MaxRetryError as e:
+            self.assertIsInstance(e.reason, SSLError)
             self.assertTrue(
-                "doesn't match" in str(e) or
-                "certificate verify failed" in str(e)
+                "doesn't match" in str(e.reason) or
+                "certificate verify failed" in str(e.reason)
             )
 
     def test_verified_with_bad_ca_certs(self):
@@ -192,10 +188,11 @@ class TestHTTPS(HTTPSDummyServerTestCase):
         try:
             https_pool.request('GET', '/')
             self.fail("Didn't raise SSL error with bad CA certs")
-        except SSLError as e:
-            self.assertTrue('certificate verify failed' in str(e),
+        except MaxRetryError as e:
+            self.assertIsInstance(e.reason, SSLError)
+            self.assertTrue('certificate verify failed' in str(e.reason),
                             "Expected 'certificate verify failed',"
-                            "instead got: %r" % e)
+                            "instead got: %r" % e.reason)
 
     def test_verified_without_ca_certs(self):
         # default is cert_reqs=None which is ssl.CERT_NONE
@@ -207,23 +204,26 @@ class TestHTTPS(HTTPSDummyServerTestCase):
             https_pool.request('GET', '/')
             self.fail("Didn't raise SSL error with no CA certs when"
                       "CERT_REQUIRED is set")
-        except SSLError as e:
+        except MaxRetryError as e:
+            self.assertIsInstance(e.reason, SSLError)
             # there is a different error message depending on whether or
             # not pyopenssl is injected
-            self.assertTrue('No root certificates specified' in str(e) or
-                            'certificate verify failed' in str(e) or
-                            'invalid certificate chain' in str(e),
+            self.assertTrue('No root certificates specified' in str(e.reason) or
+                            'certificate verify failed' in str(e.reason) or
+                            'invalid certificate chain' in str(e.reason),
                             "Expected 'No root certificates specified',  "
                             "'certificate verify failed', or "
                             "'invalid certificate chain', "
-                            "instead got: %r" % e)
+                            "instead got: %r" % e.reason)
 
     def test_no_ssl(self):
         pool = HTTPSConnectionPool(self.host, self.port)
         pool.ConnectionCls = None
         self.addCleanup(pool.close)
         self.assertRaises(SSLError, pool._new_conn)
-        self.assertRaises(SSLError, pool.request, 'GET', '/')
+        with self.assertRaises(MaxRetryError) as cm:
+            pool.request('GET', '/', retries=0)
+        self.assertIsInstance(cm.exception.reason, SSLError)
 
     def test_unverified_ssl(self):
         """ Test that bare HTTPSConnection can connect, make requests """
@@ -330,17 +330,22 @@ class TestHTTPS(HTTPSDummyServerTestCase):
         https_pool.assert_fingerprint = 'AA:AA:AA:AA:AA:AAAA:AA:AAAA:AA:' \
                                         'AA:AA:AA:AA:AA:AA:AA:AA:AA'
 
-        self.assertRaises(SSLError, https_pool.request, 'GET', '/')
+        def _test_request(pool):
+            with self.assertRaises(MaxRetryError) as cm:
+                pool.request('GET', '/', retries=0)
+            self.assertIsInstance(cm.exception.reason, SSLError)
+
+        _test_request(https_pool)
         https_pool._get_conn()
 
         # Uneven length
         https_pool.assert_fingerprint = 'AA:A'
-        self.assertRaises(SSLError, https_pool.request, 'GET', '/')
+        _test_request(https_pool)
         https_pool._get_conn()
 
         # Invalid length
         https_pool.assert_fingerprint = 'AA'
-        self.assertRaises(SSLError, https_pool.request, 'GET', '/')
+        _test_request(https_pool)
 
     def test_verify_none_and_bad_fingerprint(self):
         https_pool = HTTPSConnectionPool('127.0.0.1', self.port,
@@ -350,7 +355,9 @@ class TestHTTPS(HTTPSDummyServerTestCase):
 
         https_pool.assert_fingerprint = 'AA:AA:AA:AA:AA:AAAA:AA:AAAA:AA:' \
                                         'AA:AA:AA:AA:AA:AA:AA:AA:AA'
-        self.assertRaises(SSLError, https_pool.request, 'GET', '/')
+        with self.assertRaises(MaxRetryError) as cm:
+            https_pool.request('GET', '/', retries=0)
+        self.assertIsInstance(cm.exception.reason, SSLError)
 
     def test_verify_none_and_good_fingerprint(self):
         https_pool = HTTPSConnectionPool('127.0.0.1', self.port,
@@ -414,6 +421,7 @@ class TestHTTPS(HTTPSDummyServerTestCase):
                                          cert_reqs='CERT_NONE')
         self.addCleanup(https_pool.close)
         conn = https_pool._new_conn()
+        self.addCleanup(conn.close)
         try:
             conn.set_tunnel(self.host, self.port)
         except AttributeError:  # python 2.6
@@ -464,6 +472,7 @@ class TestHTTPS(HTTPSDummyServerTestCase):
         fingerprint = '92:81:FE:85:F7:0C:26:60:EC:D6:B3:BF:93:CF:F9:71:CC:07:7D:0A'
 
         conn = VerifiedHTTPSConnection(self.host, self.port)
+        self.addCleanup(conn.close)
         https_pool = HTTPSConnectionPool(self.host, self.port,
                                          cert_reqs='CERT_REQUIRED',
                                          ca_certs=DEFAULT_CA,
@@ -515,7 +524,9 @@ class TestHTTPS_TLSv1(HTTPSDummyServerTestCase):
         # is an issue with the OpenSSL for Python 2.6 on Windows.
 
         self._pool.cert_reqs = 'CERT_REQUIRED'
-        self.assertRaises(SSLError, self._pool.request, 'GET', '/')
+        with self.assertRaises(MaxRetryError) as cm:
+            self._pool.request('GET', '/', retries=0)
+        self.assertIsInstance(cm.exception.reason, SSLError)
         self._pool.ca_certs = DEFAULT_CA
         self._pool.request('GET', '/')